Remove redundant cast in UART transmit
Remove unnecessary (uint32_t) casts when writing transmit data for the GD32H7XX USART in lib-gd32/src/gd32_uart.cpp. The two assignments in Gd32UartTransmit and Gd32UartTransmitString now use USART_TDATA_TDATA & *data++, matching the non-H7 branch and reducing redundancy (and potential compiler warnings).

+++ lib-gd32/src/gd32_uart.cpp
@@ -309,7 +309,7 @@ void Gd32UartTransmit(uint32_t usart_periph, const uint8_t* data, uint32_t lengt
     while (length-- != 0) {
         while (RESET == usart_flag_get(usart_periph, USART_FLAG_TBE));
 #if defined(GD32H7XX)
-        USART_TDATA(usart_periph) = USART_TDATA_TDATA & (uint32_t)*data++;
+        USART_TDATA(usart_periph) = USART_TDATA_TDATA & *data++;
 #else
         USART_DATA(usart_periph) = (USART_DATA_DATA & *data++);
 #endif
@@ -324,7 +324,7 @@ void Gd32UartTransmitString(uint32_t usart_periph, const char* data) {
     while (*data != '\0') {
         while (RESET == usart_flag_get(USART0, USART_FLAG_TBE));
 #if defined(GD32H7XX)
-        USART_TDATA(usart_periph) = USART_TDATA_TDATA & (uint32_t)*data++;
+        USART_TDATA(usart_periph) = USART_TDATA_TDATA & *data++;
 #else
         USART_DATA(usart_periph) = (USART_DATA_DATA & *data++);
 #endif
